MyRect 구조체의 생성자 변경
MyRect 구조체에는 원점(origin)과 크기를 매개변수로 하는 생성자 추가.
원점은 소숫점이 될 수 있으므로, MyPoint의 x, y 프로퍼티를 Float로 변경.
이에 따른 추가 수정.

+++ CoordinateCalculator/CoordinateCalculator/InputView.swift
@@ -49,8 +49,8 @@ extension InputView {
         var pointArray = [MyPoint]()
         for point in points {
             guard let xy = splitXY(point) else { throw InputError.invalidInput }
-            let xNum: Int = Int(xy.0) ?? 0
-            let yNum: Int = Int(xy.1) ?? 0
+            let xNum: Float = Float(xy.0) ?? 0
+            let yNum: Float = Float(xy.1) ?? 0
             if xNum > 24 || yNum > 24 { throw InputError.outOfNumber }
             pointArray.append(MyPoint(x: xNum, y: yNum))
         }
